+++ client/src/hooks/impact/ranking.ts
@@ -3,8 +3,8 @@ import { useQuery } from '@tanstack/react-query';
 import { apiRawService } from 'services/api';
 
 import type { UseQueryOptions } from '@tanstack/react-query';
-import type { ImpactRanking } from 'types';
 import type { ImpactTabularAPIParams } from 'types';
+import type { ImpactDataApiResponse } from './types';
 
 const DEFAULT_QUERY_OPTIONS = {
   placeholderData: {
@@ -21,15 +21,17 @@ type ImpactRankingParams = ImpactTabularAPIParams & {
   sort: string;
 };
 
-export const useImpactRanking = <T = ImpactRanking>(
+type ResponseType = ImpactDataApiResponse<false>;
+
+export const useImpactRanking = <T = ResponseType>(
   params: Partial<ImpactRankingParams> = { maxRankingEntities: 5, sort: 'ASC' },
-  options: UseQueryOptions<ImpactRanking, unknown, T, ['impact-ranking', typeof params]> = {},
+  options: UseQueryOptions<ResponseType, unknown, T, ['impact-ranking', typeof params]> = {},
 ) => {
   const query = useQuery(
     ['impact-ranking', params],
     () =>
       apiRawService
-        .get<ImpactRanking>('/impact/ranking', {
+        .get<ResponseType>('/impact/ranking', {
           params,
         })
         .then((response) => {
