+++ client/src/types.d.ts
@@ -124,6 +124,8 @@ export type SourcingLocation = {
   businessUnit: string;
   country: string;
   locationType: string;
+  // TODO: fix this
+  // purchases: { year: number; tonnage: number }[];
   purchases: Record<string, integer>;
 };
 
@@ -171,94 +173,11 @@ type AggregatedValues = Readonly<{
   sort: 'DES' | 'ASC';
 }>;
 
-interface CommonValueItemProps {
-  name: string;
-  year: number;
-  children?: this;
-}
-
-interface CommonComparisonValueItemProps extends CommonValueItemProps {
-  absoluteDifference: number;
-  percentageDifference: number;
-  isProjected: boolean;
-}
-
-interface PlainValueItemProps extends CommonValueItemProps {
-  value: number;
-}
-
-interface ComparisonValueItemProps extends CommonComparisonValueItemProps, PlainValueItemProps {
-  scenarioValue: number;
-}
-
-interface ScenarioComparisonValueItemProps extends CommonComparisonValueItemProps {
-  scenarioOneValue: number;
-  scenarioTwoValue: number;
-}
-
-export type ImpactRowType<
-  Comparison extends TableComparisonMode,
-  IsParent extends boolean = false,
-> = IsParent extends true
-  ? ImpactTableData<Comparison>
-  : {
-      children?: this[];
-      name: string;
-      values: ImpactTableValueItem<Comparison>[];
-    };
-
-export type TableComparisonMode = boolean | 'scenario';
-
-export type ImpactTableValueItem<Comparison extends TableComparisonMode> = Comparison extends false
-  ? PlainValueItemProps
-  : Comparison extends 'scenario'
-  ? ScenarioComparisonValueItemProps
-  : ComparisonValueItemProps;
-
-export interface ImpactTableData<Comparison extends TableComparisonMode> {
-  groupBy: string;
-  indicatorId: Indicator['id'];
-  indicatorShortName: string;
-  metadata: Metadata;
-  rows: ImpactRowType<Comparison>[];
-  yearSum: Comparison extends false
-    ? Omit<ImpactTableValueItem<Comparison>, 'isProjected'>[]
-    : ImpactTableValueItem<Comparison>[];
-}
-
-export type PurchasedTonnesData = {
-  year: number;
-  value: number;
-  isProjected: boolean;
-};
-
 export type PaginationMetadata = {
   page?: number;
   size?: number;
   totalItems?: number;
   totalPages?: number;
-};
-
-export type ImpactDataApiResponse<Comparison extends TableComparisonMode> = Omit<
-  ImpactData<Comparison>,
-  'data'
-> & {
-  data: Omit<ImpactData<Comparison>['data'], 'impactTable'> & Comparison extends 'scenario'
-    ? { scenarioVsScenarioImpactTable: ImpactTableData<Comparison>[] }
-    : { impactTable: ImpactTableData<Comparison>[] };
-};
-
-export type ImpactData<Comparison extends TableComparisonMode> = {
-  data: {
-    impactTable: ImpactTableData<Comparison>[];
-    purchasedTonnes?: PurchasedTonnesData[];
-  };
-  metadata: PaginationMetadata;
-};
-
-export type ImpactRanking = {
-  impactTable: ImpactTableData[];
-  purchasedTonnes: PurchasedTonnesData[];
 };
 
 /**
